src/swellsight/evaluation/benchmarks.py
"""
Performance benchmarking system for wave analysis models.

Implements inference speed benchmarking, memory usage profiling,
and throughput testing across different hardware configurations.
"""

from typing import Dict, Any, List
import time
import torch
import psutil
import numpy as np
from dataclasses import dataclass
from contextlib import contextmanager

@dataclass
class HardwareConfig:
    """Hardware configuration information."""
    device_type: str  # "cuda" or "cpu"
    device_name: str
    memory_total_gb: float
    compute_capability: str = None  # For CUDA devices

@dataclass
class PerformanceBenchmark:
    """Performance benchmark results."""
    inference_time_ms: float
    memory_usage_mb: float
    throughput_images_per_second: float
    gpu_utilization: float
    cpu_utilization: float
    hardware_config: HardwareConfig

class PerformanceBenchmarker:
    """Performance benchmarking system for wave analysis models."""
    
    def __init__(self, model: torch.nn.Module, device: torch.device):
        """Initialize performance benchmarker.
        
        Args:
            model: Model to benchmark
            device: Device to run benchmarks on
        """
        self.model = model
        self.device = device
        self.hardware_config = self._detect_hardware_config()
        
        # Move model to device and set to eval mode
        self.model.to(self.device)
        self.model.eval()
    
    def _detect_hardware_config(self) -> HardwareConfig:
        """Detect current hardware configuration."""
        # TODO: Implement hardware detection in task 11.2
        return HardwareConfig(
            device_type=self.device.type,
            device_name="Unknown",
            memory_total_gb=0.0
        )
    
    @contextmanager
    def _measure_time(self):
        """Context manager for measuring execution time."""
        if self.device.type == "cuda":
            torch.cuda.synchronize()
        
        start_time = time.perf_counter()
        yield
        
        if self.device.type == "cuda":
            torch.cuda.synchronize()
        
        end_time = time.perf_counter()
        self.last_execution_time = (end_time - start_time) * 1000  # Convert to ms
    
    def benchmark_inference_speed(self, 
                                 input_tensor: torch.Tensor,
                                 num_warmup: int = 10,
                                 num_iterations: int = 100) -> float:
        """Benchmark model inference speed.
        
        Args:
            input_tensor: Sample input tensor
            num_warmup: Number of warmup iterations
            num_iterations: Number of benchmark iterations
            
        Returns:
            Average inference time in milliseconds
        """
        # TODO: Implement inference speed benchmarking in task 11.2
        raise NotImplementedError("Inference speed benchmarking will be implemented in task 11.2")
    
    def measure_memory_usage(self, input_tensor: torch.Tensor) -> Dict[str, float]:
        """Measure memory usage during inference.
        
        Args:
            input_tensor: Sample input tensor
            
        Returns:
            Dictionary with memory usage statistics
        """
        # TODO: Implement memory usage measurement in task 11.2
        raise NotImplementedError("Memory usage measurement will be implemented in task 11.2")
    
    def benchmark_throughput(self, 
                           batch_sizes: List[int],
                           input_shape: tuple) -> Dict[int, float]:
        """Benchmark throughput for different batch sizes.
        
        Args:
            batch_sizes: List of batch sizes to test
            input_shape: Shape of individual input (C, H, W)
            
        Returns:
            Dictionary mapping batch size to throughput (images/second)
        """
        # TODO: Implement throughput benchmarking in task 11.2
        raise NotImplementedError("Throughput benchmarking will be implemented in task 11.2")
    
    def run_complete_benchmark(self, 
                             input_tensor: torch.Tensor) -> PerformanceBenchmark:
        """Run complete performance benchmark.
        
        Args:
            input_tensor: Sample input tensor for benchmarking
            
        Returns:
            PerformanceBenchmark with comprehensive performance metrics
        """
        # Benchmark inference speed
        inference_time = self.benchmark_inference_speed(input_tensor)
        
        # Measure memory usage
        memory_stats = self.measure_memory_usage(input_tensor)
        
        # Calculate throughput
        throughput = 1000.0 / inference_time  # images per second
        
        return PerformanceBenchmark(
            inference_time_ms=inference_time,
            memory_usage_mb=memory_stats.get("peak_memory_mb", 0),
            throughput_images_per_second=throughput,
            gpu_utilization=memory_stats.get("gpu_utilization", 0),
            cpu_utilization=psutil.cpu_percent(),
            hardware_config=self.hardware_config
        )