src/swellsight/evaluation/metrics.py
"""
Wave analysis metrics for model evaluation.

Implements accuracy metrics for wave height, direction classification,
and breaking type classification with confidence calibration.
"""

from typing import Dict, List, Tuple
import numpy as np
from dataclasses import dataclass
from sklearn.metrics import accuracy_score, precision_recall_fscore_support, confusion_matrix

@dataclass
class HeightAccuracyMetrics:
    """Accuracy metrics for wave height prediction."""
    mae: float  # Mean Absolute Error
    rmse: float  # Root Mean Square Error
    accuracy_within_02m: float  # Percentage within ±0.2m
    accuracy_within_05m: float  # Percentage within ±0.5m
    extreme_condition_detection: float

@dataclass
class ClassificationMetrics:
    """Metrics for classification tasks (direction, breaking type)."""
    accuracy: float
    precision_per_class: Dict[str, float]
    recall_per_class: Dict[str, float]
    f1_score_per_class: Dict[str, float]
    confusion_matrix: np.ndarray
    macro_avg_f1: float

@dataclass
class CalibrationMetrics:
    """Confidence calibration metrics."""
    expected_calibration_error: float
    maximum_calibration_error: float
    reliability_diagram_data: Dict[str, List[float]]

@dataclass
class AccuracyMetrics:
    """Complete accuracy metrics for all wave analysis tasks."""
    height_metrics: HeightAccuracyMetrics
    direction_metrics: ClassificationMetrics
    breaking_type_metrics: ClassificationMetrics
    confidence_calibration: CalibrationMetrics
    overall_score: float

class WaveAnalysisMetrics:
    """Comprehensive metrics calculator for wave analysis evaluation."""
    
    def __init__(self):
        """Initialize wave analysis metrics calculator."""
        self.class_names = {
            "direction": ["LEFT", "RIGHT", "STRAIGHT"],
            "breaking": ["SPILLING", "PLUNGING", "SURGING"]
        }
    
    def calculate_height_metrics(self, 
                                predictions: np.ndarray,
                                targets: np.ndarray) -> HeightAccuracyMetrics:
        """Calculate wave height accuracy metrics.
        
        Args:
            predictions: Predicted wave heights in meters
            targets: Ground truth wave heights in meters
            
        Returns:
            HeightAccuracyMetrics with detailed accuracy assessment
        """
        # TODO: Implement height metrics in task 11.1
        raise NotImplementedError("Height metrics will be implemented in task 11.1")
    
    def calculate_classification_metrics(self,
                                       predictions: np.ndarray,
                                       targets: np.ndarray,
                                       task_type: str) -> ClassificationMetrics:
        """Calculate classification metrics for direction or breaking type.
        
        Args:
            predictions: Predicted class indices
            targets: Ground truth class indices
            task_type: "direction" or "breaking"
            
        Returns:
            ClassificationMetrics with detailed classification assessment
        """
        # TODO: Implement classification metrics in task 11.1
        raise NotImplementedError("Classification metrics will be implemented in task 11.1")
    
    def calculate_calibration_metrics(self,
                                    confidences: np.ndarray,
                                    accuracies: np.ndarray) -> CalibrationMetrics:
        """Calculate confidence calibration metrics.
        
        Args:
            confidences: Predicted confidence scores
            accuracies: Binary accuracy indicators
            
        Returns:
            CalibrationMetrics with calibration assessment
        """
        # TODO: Implement calibration metrics in task 11.1
        raise NotImplementedError("Calibration metrics will be implemented in task 11.1")
    
    def calculate_complete_metrics(self,
                                 height_preds: np.ndarray,
                                 height_targets: np.ndarray,
                                 direction_preds: np.ndarray,
                                 direction_targets: np.ndarray,
                                 breaking_preds: np.ndarray,
                                 breaking_targets: np.ndarray,
                                 confidences: Dict[str, np.ndarray]) -> AccuracyMetrics:
        """Calculate complete metrics for all wave analysis tasks.
        
        Args:
            height_preds: Wave height predictions
            height_targets: Wave height ground truth
            direction_preds: Direction predictions
            direction_targets: Direction ground truth
            breaking_preds: Breaking type predictions
            breaking_targets: Breaking type ground truth
            confidences: Confidence scores for all tasks
            
        Returns:
            AccuracyMetrics with comprehensive evaluation
        """
        # Calculate individual task metrics
        height_metrics = self.calculate_height_metrics(height_preds, height_targets)
        direction_metrics = self.calculate_classification_metrics(
            direction_preds, direction_targets, "direction"
        )
        breaking_metrics = self.calculate_classification_metrics(
            breaking_preds, breaking_targets, "breaking"
        )
        
        # Calculate confidence calibration
        # TODO: Implement proper calibration calculation
        calibration_metrics = CalibrationMetrics(
            expected_calibration_error=0.0,
            maximum_calibration_error=0.0,
            reliability_diagram_data={}
        )
        
        # Calculate overall score
        overall_score = (
            height_metrics.accuracy_within_02m * 0.4 +
            direction_metrics.accuracy * 0.3 +
            breaking_metrics.accuracy * 0.3
        )
        
        return AccuracyMetrics(
            height_metrics=height_metrics,
            direction_metrics=direction_metrics,
            breaking_type_metrics=breaking_metrics,
            confidence_calibration=calibration_metrics,
            overall_score=overall_score
        )